tests: Move `fn setup` to fixture module
