+++ tests/Cargo.toml
@@ -9,18 +9,18 @@ anyhow = "1.0"
 futures = { version = "0.3", features = ["compat"] }
 graph = { path = "../graph" }
 tokio = {version = "1.16.1", features = ["rt", "macros", "process"]}
-
-[dev-dependencies]
-slog = { version = "2.7.0", features = ["release_max_level_trace", "max_level_trace"] }
-bollard = "0.10"
-anyhow = "1.0.57"
-lazy_static = "1.4.0"
-tokio-stream = "0.1"
+graph-chain-ethereum = { path = "../chain/ethereum" }
+async-stream = "0.3.3"
 graph-node = { path = "../node" }
 graph-core = { path = "../core" }
 graph-mock = { path = "../mock" }
 graph-store-postgres = { path = "../store/postgres" }
+slog = { version = "2.7.0", features = ["release_max_level_trace", "max_level_trace"] }
+
+[dev-dependencies]
+bollard = "0.10"
+anyhow = "1.0.57"
+lazy_static = "1.4.0"
+tokio-stream = "0.1"
 serde_yaml = "0.8"
-graph-chain-ethereum = { path = "../chain/ethereum" }
 graph-chain-near = { path = "../chain/near" }
-async-stream = "0.3.3"
